BUG: Fixed quote style
Many of the comments in the code was using Latex quote style,
which uses `` to signify a left double quote and '' for a right
double quote.

In certain circumstances, the left ones, ``, can confuse Doxygen.
In particular if there are two occurences that happen in one line
or in two successive lines, Doxygen can interpret a span of text
as an in-line code span.  This situation was occuring in the
SimpleITK ZeroCrossingImageFilter documentation, which was pulled
from ITK's ZeroCrossingImageFilter class.

To change `` and '' to ", the following sed command was run:
  sed 's/\(``\)\([ -~]*\)\('\'\''\)/"\2"/g'

Change-Id: I0f30100d5588ee07be9689a8c576fb0cd59fe16b